requirements.txt
flask
fastapi
uvicorn
opencv-python
numpy
torch
torchvision
pillow
python-dotenv